+++ dandi/delete.py
@@ -6,11 +6,11 @@ from typing import Iterable, Iterator, NamedTuple, Optional, Set, Tuple
 import click
 import requests
 
-from .consts import known_instances
+from .consts import dandiset_metadata_file
 from .dandiapi import DandiAPIClient
 from .dandiarchive import parse_dandi_url
 from .exceptions import NotFoundError
-from .utils import is_url
+from .utils import get_instance, is_url
 
 
 class RemoteAsset(NamedTuple):
@@ -37,7 +37,9 @@ class Deleter:
 
     def set_dandiset(self, api_url: str, dandiset_id: str) -> None:
         if self.client is None:
-            self.client = DandiAPIClient(api_url)
+            # Strip the trailing slash so that dandi_authenticate can find the
+            # URL in known_instances_rev:
+            self.client = DandiAPIClient(api_url.rstrip("/"))
             self.client.dandi_authenticate()
             self.dandiset_id = dandiset_id
             try:
@@ -113,7 +115,7 @@ class Deleter:
             raise RuntimeError(f"Unexpected asset type for {url}: {asset_type}")
 
     def register_local_path_equivalent(self, instance_name: str, filepath: str) -> None:
-        instance = known_instances[instance_name]
+        instance = get_instance(instance_name)
         if instance.girder is not None:
             raise NotImplementedError("Cannot delete assets from Girder instances")
         api_url = instance.api
@@ -210,6 +212,11 @@ def find_local_asset(filepath: str) -> Tuple[str, str]:
 
     path = Path(filepath).absolute()
     dandiset = Dandiset.find(path.parent)
+    if dandiset is None:
+        raise RuntimeError(
+            f"Found no {dandiset_metadata_file} anywhere.  "
+            "Use 'dandi register', 'download', or 'organize' first"
+        )
     relpath = str(PurePosixPath(path.relative_to(dandiset.path)))
     if path.is_dir():
         relpath += "/"