+++ dandi/validate.py
@@ -1,10 +1,13 @@
 import os.path as op
 
 from .consts import dandiset_metadata_file
+from . import get_logger
 from .metadata import get_metadata
 from .pynwb_utils import validate as pynwb_validate
 from .pynwb_utils import validate_cache
 from .utils import find_dandi_files, yaml_load
+
+lgr = get_logger()
 
 # TODO -- should come from schema.  This is just a simplistic example for now
 _required_dandiset_metadata_fields = ["identifier", "name", "description"]
@@ -69,10 +72,19 @@ def validate_dandiset_yaml(filepath, schema_version=None, devel_debug=False):
         except ValidationError as e:
             if devel_debug:
                 raise
+            lgr.warning(
+                "Validation error for %s: %s", filepath, e, extra={"validating": True}
+            )
             return [str(e)]
         except Exception as e:
             if devel_debug:
                 raise
+            lgr.warning(
+                "Unexpected validation error for %s: %s",
+                filepath,
+                e,
+                extra={"validating": True},
+            )
             return [f"Failed to convert metadata: {e}"]
         return []
 
@@ -97,10 +109,19 @@ def validate_dandi_nwb(filepath, schema_version=None, devel_debug=False):
         except ValidationError as e:
             if devel_debug:
                 raise
+            lgr.warning(
+                "Validation error for %s: %s", filepath, e, extra={"validating": True}
+            )
             return [str(e)]
         except Exception as e:
             if devel_debug:
                 raise
+            lgr.warning(
+                "Unexpected validation error for %s: %s",
+                filepath,
+                e,
+                extra={"validating": True},
+            )
             return [f"Failed to read metadata: {e}"]
         return []
     else:
@@ -110,6 +131,12 @@ def validate_dandi_nwb(filepath, schema_version=None, devel_debug=False):
         except Exception as e:
             if devel_debug:
                 raise
+            lgr.warning(
+                "Failed to read metadata in %s: %s",
+                filepath,
+                e,
+                extra={"validating": True},
+            )
             return [f"Failed to read metadata: {e}"]
         return _check_required_fields(meta, _required_nwb_metadata_fields)